Update Film MCD Commit

+++ database/migrations/2016_10_013_054747_create_film_table.php
@@ -17,11 +17,11 @@ class CreateFilmTable extends Migration
             $table->smallInteger('language_id');
             $table->string('title', 255);
             $table->text('description');
-            $table->smallInteger('release_year');
+            $table->tinyInteger('release_year');
             $table->smallInteger('original_language_id');
             $table->tinyInteger('rental_duration');
             $table->decimal('rental_rate', 4, 2);
-            $table->smallInteger('length');
+            $table->tinyInteger('length');
             $table->decimal('replacement_cost', 5, 2);
             $table->enum('rating', array("G","PG","PG-1","R","NC-17",));
             $table->addColumn('set', 'special_features', array("Trailers","Commentaries","Deleted Scenes","Behind The Scenes",));
